requirements.txt
streamlit==1.41.1
pypdf==5.1.0
numpy==2.1.3
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
google-generativeai==0.8.3
python-dotenv==1.0.1
reportlab==4.2.5
